Moves xmlutils unit tests into src xmlutil.rs file instead of being in integration test folder.
